fix: fixed project compilation issue
