+++ common/src/main/java/ru/practicum/common/model/Item.java
@@ -1,11 +1,10 @@
-package ru.practicum.server.item.model;
+package ru.practicum.common.model;
 
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import lombok.*;
 import lombok.experimental.FieldDefaults;
-import ru.practicum.server.user.model.User;
 
 @FieldDefaults(level = AccessLevel.PRIVATE)
 @Entity
