+++ common/src/main/java/ru/practicum/common/model/ItemRequest.java
@@ -1,13 +1,14 @@
-package ru.practicum.server.request.model;
+package ru.practicum.common.model;
 
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotBlank;
 import lombok.*;
 import lombok.experimental.FieldDefaults;
 
+import java.time.LocalDateTime;
+import java.time.ZoneId;
+import java.time.ZoneOffset;
 import java.time.ZonedDateTime;
-
-import static ru.practicum.server.booking.mapper.BookingMapper.getUtcNow;
 
 @FieldDefaults(level = AccessLevel.PRIVATE)
 @Entity
@@ -32,4 +33,16 @@ public class ItemRequest {
     @Builder.Default
     @Column
     ZonedDateTime created = getUtcNow();
+
+    public static LocalDateTime utcToLocal(ZonedDateTime zonedTime) {
+        return zonedTime.withZoneSameInstant(ZoneId.systemDefault()).toLocalDateTime();
+    }
+
+    public static ZonedDateTime localToUtc(LocalDateTime localTime) {
+        return localTime.atZone(ZoneId.systemDefault()).withZoneSameInstant(ZoneOffset.UTC);
+    }
+
+    public static ZonedDateTime getUtcNow() {
+        return localToUtc(LocalDateTime.now());
+    }
 }
