+++ common/src/main/java/ru/practicum/common/model/User.java
@@ -1,4 +1,4 @@
-package ru.practicum.server.user.model;
+package ru.practicum.common.model;
 
 import jakarta.persistence.*;
 import jakarta.validation.constraints.Email;
