+++ gateway/src/main/java/ru/practicum/gateway/ShareItGateway.java
@@ -1,4 +1,4 @@
-package ru.gateway;
+package ru.practicum.gateway;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
