+++ gateway/src/main/java/ru/practicum/gateway/client/BaseClient.java
@@ -1,4 +1,4 @@
-package ru.gateway.client;
+package ru.practicum.gateway.client;
 
 import java.util.List;
 import java.util.Map;
