+++ gateway/src/main/java/ru/practicum/gateway/util/ErrorHandler.java
@@ -1,4 +1,4 @@
-package ru.gateway.util;
+package ru.practicum.gateway.util;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.MethodArgumentNotValidException;
