+++ gateway/src/main/java/ru/practicum/gateway/util/ErrorResponse.java
@@ -1,4 +1,4 @@
-package ru.gateway.util;
+package ru.practicum.gateway.util;
 
 public record ErrorResponse(String description, String error) {
 }
