--- gateway/src/test/java/ru/gateway/ShareItTests.java
@@ -1,13 +1,0 @@
-package ru.gateway;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ShareItTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
