+++ server/pom.xml
@@ -11,7 +11,7 @@
     <artifactId>shareit-server</artifactId>
     <version>0.0.1-SNAPSHOT</version>
 
-    <name>ShareIt Server</name>
+<!--    <name>ShareIt Server</name>-->
 
     <dependencies>
         <dependency>
