+++ server/src/main/java/ru/practicum/server/ShareItServer.java
@@ -1,0 +1,9 @@
+package ru.practicum.server;
+
+import org.springframework.boot.SpringApplication;
+
+public class ShareItServer {
+    public static void main(String[] args) {
+        SpringApplication.run(ShareItServer.class, args);
+    }
+}
