+++ server/src/main/java/ru/practicum/server/booking/mapper/BookingMapper.java
@@ -2,19 +2,17 @@ package ru.practicum.server.booking.mapper;
 
 import ru.practicum.common.dto.booking.BookingDtoCreation;
 import ru.practicum.common.dto.booking.BookingDtoResponse;
-import ru.practicum.server.booking.model.Booking;
+import ru.practicum.common.model.Booking;
 import ru.practicum.common.dto.booking.BookingStatus;
 import ru.practicum.common.dto.item.ItemDto;
 import ru.practicum.server.item.mapper.ItemMapper;
-import ru.practicum.server.item.model.Item;
+import ru.practicum.common.model.Item;
 import ru.practicum.common.dto.user.UserDto;
 import ru.practicum.server.user.mapper.UserMapper;
-import ru.practicum.server.user.model.User;
+import ru.practicum.common.model.User;
 
-import java.time.LocalDateTime;
-import java.time.ZoneId;
-import java.time.ZoneOffset;
-import java.time.ZonedDateTime;
+import static ru.practicum.common.model.ItemRequest.localToUtc;
+import static ru.practicum.common.model.ItemRequest.utcToLocal;
 
 
 public class BookingMapper {
@@ -30,17 +28,5 @@ public class BookingMapper {
     public static Booking dtoToNewBooking(BookingDtoCreation dto, Item item, User booker) {
         return new Booking(null, item, booker, localToUtc(dto.getStart()),
                 localToUtc(dto.getEnd()), BookingStatus.WAITING);
-    }
-
-    public static LocalDateTime utcToLocal(ZonedDateTime zonedTime) {
-        return zonedTime.withZoneSameInstant(ZoneId.systemDefault()).toLocalDateTime();
-    }
-
-    public static ZonedDateTime localToUtc(LocalDateTime localTime) {
-        return localTime.atZone(ZoneId.systemDefault()).withZoneSameInstant(ZoneOffset.UTC);
-    }
-
-    public static ZonedDateTime getUtcNow() {
-        return localToUtc(LocalDateTime.now());
     }
 }
