+++ server/src/main/java/ru/practicum/server/item/mapper/CommentMapper.java
@@ -1,19 +1,19 @@
 package ru.practicum.server.item.mapper;
 
-import ru.practicum.server.booking.mapper.BookingMapper;
 import ru.practicum.common.dto.item.CommentDto;
-import ru.practicum.server.item.model.Comment;
-import ru.practicum.server.item.model.Item;
-import ru.practicum.server.user.model.User;
+import ru.practicum.common.model.Comment;
+import ru.practicum.common.model.Item;
+import ru.practicum.common.model.User;
 
 import java.util.List;
 
-import static ru.practicum.server.booking.mapper.BookingMapper.getUtcNow;
+import static ru.practicum.common.model.ItemRequest.getUtcNow;
+import static ru.practicum.common.model.ItemRequest.utcToLocal;
 
 public class CommentMapper {
     public static CommentDto commentToDto(Comment comment) {
         return new CommentDto(comment.getId(), comment.getCommentText(),
-                comment.getAuthor().getName(), BookingMapper.utcToLocal(comment.getCreated()));
+                comment.getAuthor().getName(), utcToLocal(comment.getCreated()));
     }
 
     public static List<CommentDto> commentToDto(List<Comment> comments) {
