+++ server/src/main/java/ru/practicum/server/item/service/ItemServiceImpl.java
@@ -4,7 +4,7 @@ import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
 import lombok.experimental.FieldDefaults;
 import org.springframework.stereotype.Service;
-import ru.practicum.server.booking.model.Booking;
+import ru.practicum.common.model.Booking;
 import ru.practicum.common.dto.booking.BookingStatus;
 import ru.practicum.server.booking.repository.BookingRepository;
 import ru.practicum.common.exception.NotFoundException;
@@ -12,17 +12,17 @@ import ru.practicum.common.exception.ValidationException;
 import ru.practicum.common.dto.item.CommentDto;
 import ru.practicum.common.dto.item.ItemDto;
 import ru.practicum.server.item.mapper.CommentMapper;
-import ru.practicum.server.item.model.Comment;
+import ru.practicum.common.model.Comment;
 import ru.practicum.server.item.repository.ItemRepository;
 import ru.practicum.server.item.repository.CommentRepository;
 import ru.practicum.common.dto.item.ItemDtoExtended;
 import ru.practicum.server.item.mapper.ItemMapper;
-import ru.practicum.server.item.model.Item;
+import ru.practicum.common.model.Item;
 import ru.practicum.server.request.repository.ItemRequestRepository;
-import ru.practicum.server.user.model.User;
+import ru.practicum.common.model.User;
 import ru.practicum.server.user.repository.UserRepository;
 
-import static ru.practicum.server.booking.mapper.BookingMapper.getUtcNow;
+import static ru.practicum.common.model.ItemRequest.getUtcNow;
 import static ru.practicum.server.request.service.ItemRequestServiceImpl.REQUEST_NOT_FOUND;
 import static ru.practicum.server.user.service.UserServiceImpl.USER_NOT_FOUND;
