+++ server/src/main/java/ru/practicum/server/request/service/ItemRequestServiceImpl.java
@@ -5,10 +5,10 @@ import lombok.RequiredArgsConstructor;
 import lombok.experimental.FieldDefaults;
 import org.springframework.stereotype.Service;
 import ru.practicum.common.exception.NotFoundException;
-import ru.practicum.server.item.model.Item;
+import ru.practicum.common.model.Item;
 import ru.practicum.server.item.repository.ItemRepository;
 import ru.practicum.common.dto.request.ItemRequestDtoExtended;
-import ru.practicum.server.request.model.ItemRequest;
+import ru.practicum.common.model.ItemRequest;
 import ru.practicum.common.dto.request.ItemRequestDto;
 import ru.practicum.server.request.mapper.ItemRequestMapper;
 import ru.practicum.server.request.repository.ItemRequestRepository;
